+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/content/MediaAttachmentContent.kt
@@ -115,7 +115,7 @@ public fun MediaAttachmentContent(
         }
     },
 ) {
-    val (message, onLongItemClick, onMediaGalleryPreviewResult) = attachmentState
+    val (message, isMine, onLongItemClick, onMediaGalleryPreviewResult) = attachmentState
     val gridSpacing = ChatTheme.dimens.attachmentsContentMediaGridSpacing
 
     Row(